python-interactive/requirements.txt
# requirements.txt
pika
requests
pandas
numpy
google-auth
google-cloud-core
google-cloud-vision
Pillow
# psycopg2-binary
# sqlalchemy
minio
# pyzbar
# opencv-python
scikit-image
jupyter
